Don't blow up at plugin start up due to the plugin path not being ready yet
CL-555
